+++ abdm_integrator/abha/utils/decorators.py
@@ -1,7 +1,7 @@
 from functools import wraps
 from typing import List
 
-from abdm_python_integrator.abha.utils.response import get_bad_response
+from abdm_integrator.abha.utils.response import get_bad_response
 
 
 def required_request_params(required_request_data):
