+++ abdm_integrator/abha/views/abha_creation.py
@@ -1,12 +1,12 @@
 from django.utils.decorators import method_decorator
 
-from abdm_python_integrator.abha.exceptions import INVALID_AADHAAR_MESSAGE, INVALID_MOBILE_MESSAGE
-from abdm_python_integrator.abha.utils import abha_creation as abdm_util
-from abdm_python_integrator.abha.utils.abha_creation import validate_aadhaar_number, validate_mobile_number
-from abdm_python_integrator.abha.utils.decorators import required_request_params
-from abdm_python_integrator.abha.utils.response import get_bad_response, parse_response
-from abdm_python_integrator.abha.views.base import ABHABaseView
-from abdm_python_integrator.settings import app_settings
+from abdm_integrator.abha.exceptions import INVALID_AADHAAR_MESSAGE, INVALID_MOBILE_MESSAGE
+from abdm_integrator.abha.utils import abha_creation as abdm_util
+from abdm_integrator.abha.utils.abha_creation import validate_aadhaar_number, validate_mobile_number
+from abdm_integrator.abha.utils.decorators import required_request_params
+from abdm_integrator.abha.utils.response import get_bad_response, parse_response
+from abdm_integrator.abha.views.base import ABHABaseView
+from abdm_integrator.settings import app_settings
 
 
 class GenerateAadhaarOTP(ABHABaseView):
