+++ abdm_integrator/abha/views/abha_verification.py
@@ -1,12 +1,12 @@
 from django.utils.decorators import method_decorator
 
-from abdm_python_integrator.abha.exceptions import ABHA_ERROR_MESSAGES, ABHA_IN_USE_ERROR_CODE
-from abdm_python_integrator.abha.utils import abha_verification as abdm_util
-from abdm_python_integrator.abha.utils.abha_verification import get_account_information
-from abdm_python_integrator.abha.utils.decorators import required_request_params
-from abdm_python_integrator.abha.utils.response import get_bad_response, parse_response
-from abdm_python_integrator.abha.views.base import ABHABaseView
-from abdm_python_integrator.settings import app_settings
+from abdm_integrator.abha.exceptions import ABHA_ERROR_MESSAGES, ABHA_IN_USE_ERROR_CODE
+from abdm_integrator.abha.utils import abha_verification as abdm_util
+from abdm_integrator.abha.utils.abha_verification import get_account_information
+from abdm_integrator.abha.utils.decorators import required_request_params
+from abdm_integrator.abha.utils.response import get_bad_response, parse_response
+from abdm_integrator.abha.views.base import ABHABaseView
+from abdm_integrator.settings import app_settings
 
 
 class GetAuthMethods(ABHABaseView):
