+++ abdm_integrator/exceptions.py
