+++ abdm_integrator/tests/test_utils.py
@@ -3,8 +3,8 @@ from unittest.mock import Mock, patch
 import requests
 from django.test import SimpleTestCase
 
-from abdm_python_integrator.exceptions import ABDMAccessTokenException
-from abdm_python_integrator.utils import ABDMRequestHelper
+from abdm_integrator.exceptions import ABDMAccessTokenException
+from abdm_integrator.utils import ABDMRequestHelper
 
 
 class TestABDMRequestHelper(SimpleTestCase):
@@ -20,48 +20,48 @@ class TestABDMRequestHelper(SimpleTestCase):
         mock_response.json = Mock(return_value=json_response)
         return mock_response
 
-    @patch('abdm_python_integrator.utils.requests.post')
+    @patch('abdm_integrator.utils.requests.post')
     def test_get_access_token_success(self, mocked_post):
         mocked_post.return_value = self._mock_response(json_response={'accessToken': 'test'})
         token = ABDMRequestHelper().get_access_token()
         self.assertEqual(token, 'test')
 
-    @patch('abdm_python_integrator.utils.requests.post')
+    @patch('abdm_integrator.utils.requests.post')
     def test_get_access_token_failure(self, mocked_post):
         mocked_post.return_value = self._mock_response(status_code=400)
         with self.assertRaises(ABDMAccessTokenException):
             ABDMRequestHelper().get_access_token()
 
-    @patch('abdm_python_integrator.utils.ABDMRequestHelper.get_access_token')
-    @patch('abdm_python_integrator.utils.requests.get')
+    @patch('abdm_integrator.utils.ABDMRequestHelper.get_access_token')
+    @patch('abdm_integrator.utils.requests.get')
     def test_abha_get_success(self, mocked_get, *args):
         mocked_get.return_value = self._mock_response(json_response=self.sample_success_json)
         actual_json_resp = ABDMRequestHelper().abha_get(api_path='')
         self.assertEqual(actual_json_resp, self.sample_success_json)
 
-    @patch('abdm_python_integrator.utils.ABDMRequestHelper.get_access_token')
-    @patch('abdm_python_integrator.utils.requests.get')
+    @patch('abdm_integrator.utils.ABDMRequestHelper.get_access_token')
+    @patch('abdm_integrator.utils.requests.get')
     def test_abha_get_failure(self, mocked_get,  *args):
         mocked_get.return_value = self._mock_response(status_code=400)
         with self.assertRaises(requests.HTTPError):
             ABDMRequestHelper().abha_get(api_path='')
 
-    @patch('abdm_python_integrator.utils.ABDMRequestHelper.get_access_token')
-    @patch('abdm_python_integrator.utils.requests.post')
+    @patch('abdm_integrator.utils.ABDMRequestHelper.get_access_token')
+    @patch('abdm_integrator.utils.requests.post')
     def test_abha_post_success(self, mocked_post, *args):
         mocked_post.return_value = self._mock_response(json_response=self.sample_success_json)
         actual_json_resp = ABDMRequestHelper().abha_post(api_path='', payload={})
         self.assertEqual(actual_json_resp, self.sample_success_json)
 
-    @patch('abdm_python_integrator.utils.ABDMRequestHelper.get_access_token')
-    @patch('abdm_python_integrator.utils.requests.post')
+    @patch('abdm_integrator.utils.ABDMRequestHelper.get_access_token')
+    @patch('abdm_integrator.utils.requests.post')
     def test_abha_post_failure(self, mocked_post, *args):
         mocked_post.return_value = self._mock_response(status_code=400)
         with self.assertRaises(requests.HTTPError):
             ABDMRequestHelper().abha_post(api_path='', payload={})
 
-    @patch('abdm_python_integrator.utils.ABDMRequestHelper.get_access_token')
-    @patch('abdm_python_integrator.utils.requests.post')
+    @patch('abdm_integrator.utils.ABDMRequestHelper.get_access_token')
+    @patch('abdm_integrator.utils.requests.post')
     def test_gateway_post_success(self, mocked_post, *args):
         mock_response = self._mock_response(json_response=self.sample_success_json)
         mock_response.headers = {'content-type': 'application/json'}
@@ -69,8 +69,8 @@ class TestABDMRequestHelper(SimpleTestCase):
         actual_json_resp = ABDMRequestHelper().gateway_post(api_path='', payload={})
         self.assertEqual(actual_json_resp, self.sample_success_json)
 
-    @patch('abdm_python_integrator.utils.ABDMRequestHelper.get_access_token')
-    @patch('abdm_python_integrator.utils.requests.post')
+    @patch('abdm_integrator.utils.ABDMRequestHelper.get_access_token')
+    @patch('abdm_integrator.utils.requests.post')
     def test_gateway_post_failure(self, mocked_post, *args):
         mocked_post.return_value = self._mock_response(status_code=400)
         with self.assertRaises(requests.HTTPError):
