+++ abdm_integrator/urls.py
@@ -1,0 +1,3 @@
+from abdm_integrator.abha.urls import abha_urls
+
+urlpatterns = abha_urls
