--- abdm_python_integrator/urls.py
@@ -1,3 +1,0 @@
-from abdm_python_integrator.abha.urls import abha_urls
-
-urlpatterns = abha_urls
